requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
pyjwt==2.8.0
bcrypt==4.1.2
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0